Don't convert location to point unnecessarily.
The previous version of this function would always convert the location
to point before doing the comparison. When the location was to a point
that is greater than the size of the current buffer, it would cause a
fatal error. The new version will do the conversion of location to point
only if the current buffer's file name and the other file name are a
match.