Performance fix: tsserver lookup
This is an opinionated rewrite on top of the work done by @leo-shopify.

I've kept the following from previous work:

- idea of possible sources
- idea of the order those sources should be considered

The changes consists of the following simplifications:

1. don't ever invoke npm. It's too slow.
2. instead use hardcoded logic encoded in pure elisp based assumptions
   about how npm works to resolve paths. This should be fast everywhere.

The assumptions are as follows:

1. assume package-local folder will always be called "node_modules"
   and be a direct sibling of the folder with package.json, and look
   for this folder up from the current buffer's directory.
   (I'm pretty sure this assumption holds 100%)
2. make similar fixed assumptions about global packages, but leave
   room for OS-specific locations.

Cases this approach should handle gracefully:

1. tide configured to go against specific tsserver.
2. Packages with TypeScript bundled.
3. LERNA style projects managed by projectile.
4. Loading tsserver from a *uncustomized* global package-store.
5. Bundled tsserver in case nothing else can be found.

The only case I can see I'm breaking (and that's only one) is case 4,
with a customized global package-store.

And if you *really* want to use that over any of the other
options... you can already customize tide to do that anyway, through
`tide-tsserver-executable`.

I think this is one of those cases where 90% is more than good enough
and where going for 100% is only going to give you trouble.