Sort completions using `sortText' property
`sortText' property of a completion is declared by the tsserver protocol for
use for comparing completion items so that they can be ordered.

This property has a value of "0" in most cases, but for the entries
returned by `getJavaScriptCompletionEntries()' function it's set to "1",
since these entries are intended to go after all others.