Change the cleanup implementation strategy
Instead of running it immediately when a buffer is killed, do that in an
idle timer -- but still, only once, after a buffer is killed.  Adjusting
the timer can be useful for people who only want to clean things up
every night.  Even with a short delay it's still useful in case you open
one file at a time in a directory, close it and open another, etc -- in
this case it avoids repeating the work of killing a server and starting
a new one multiple times.

Adds an `tide-project-cleanup-delay` option to control the delay,
possibly set to `nil` to disable all of this, and make
`tide-cleanup-dead-projects` interactive so it's easy to use.