Fix tide failure in magit buffers.
When a user uses magit to extract old commits of files, magit creates buffers
for these commits. These buffers are not backed by actual files. Tide is turned
on in these buffers. Prior to this commit tide would report errors whenever it
tried to contact `tsserver` to work on one of these buffers. This commit makes
it so that tide detects that it is running in a magit buffer and aborts.