Make no-server errors less naggy
Every once in a while I run into a situation where the server is either
dead or just misbehaving for whatever reason.  Since tide tends to send
a lot of messages to the server, this leads to an annoying state where
it keeps barfing over and over again.

This avoids throwing the no server error if it was just thrown.