Fix handling of timestamps in new `(TICKS . HZ)` format
Emacs is rolling out a new timestamp format, to be used by
time-handling functions such as `current-time` and `time-subtract`.
Tide assumes timestamps are in the old `(HI LO US PS)` format, and
attempts to extract a number of seconds by directly pulling out an
element from this list with `cadr`.  In order to properly work with
the changing timestamp format, we need to use Emacs' provided
functions for handling timestamps, rather than poking at the internals
directly.